+++ R/dataDoc.R
@@ -21,13 +21,13 @@
 ##' The variables are as follows:
 ##'
 ##' \itemize{
-##'   \item ID.  Person identification variable
-##'   \item sex.  Sex of the ID: 1 is female; 0 is male
-##'   \item dadID.  ID of the father
-##'   \item momID.  ID of the mother
-##'   \item FamID.  ID of the extended family
-##'   \item Gen.  Generation of the person
-##'   \item proband.  Always FALSE
+##'   \item \code{ID}.  Person identification variable
+##'   \item \code{sex}.  Sex of the ID: 1 is female; 0 is male
+##'   \item \code{dadID}.  ID of the father
+##'   \item \code{momID}.  ID of the mother
+##'   \item \code{FamID}.  ID of the extended family
+##'   \item \code{Gen}.  Generation of the person
+##'   \item \code{proband}.  Always FALSE
 ##' }
 ##'
 ##' @docType data
@@ -45,20 +45,20 @@ NULL
 ##' The variables are as follows:
 ##'
 ##' \itemize{
-##'   \item FamID.  ID of the extended family
-##'   \item ID.  Person identification variable
-##'   \item sex.  Sex of the ID: 1 is female; 0 is male
-##'   \item dadID.  ID of the father
-##'   \item momID.  ID of the mother
-##'   \item affected.  logical.  Whether the person is affected or not
-##'   \item DA1.  Binary variable signifying the meaninglessness of life
-##'   \item DA2.  Binary variable signifying the fundamental unknowability of existence
-##'   \item birthYr.  Birth year for person
-##'   \item onsetYr.  Year of onset for person
-##'   \item deathYr.  Death year for person
-##'   \item available.  logical.  Whether
-##'   \item Gen.  Generation of the person
-##'   \item proband.  logical.  Whether the person is a proband or not
+##'   \item \code{FamID}.  ID of the extended family
+##'   \item \code{ID}.  Person identification variable
+##'   \item \code{sex}.  Sex of the ID: 1 is female; 0 is male
+##'   \item \code{dadID}.  ID of the father
+##'   \item \code{momID}.  ID of the mother
+##'   \item \code{affected}.  logical.  Whether the person is affected or not
+##'   \item \code{DA1}.  Binary variable signifying the meaninglessness of life
+##'   \item \code{DA2}.  Binary variable signifying the fundamental unknowability of existence
+##'   \item \code{birthYr}.  Birth year for person
+##'   \item \code{onsetYr}.  Year of onset for person
+##'   \item \code{deathYr}.  Death year for person
+##'   \item \code{available}.  logical.  Whether
+##'   \item \code{Gen}.  Generation of the person
+##'   \item \code{proband}.  logical.  Whether the person is a proband or not
 ##' }
 ##'
 ##' @docType data
@@ -75,14 +75,14 @@ NULL
 ##' The variables are as follows:
 ##'
 ##' \itemize{
-##'   \item personID.  Person identification variable
-##'   \item famID. Family identification variable
-##'   \item name.  Name of the person
-##'   \item gen. Generation of the person
-##'   \item momID.  ID of the mother
-##'   \item dadID.  ID of the father
-##'   \item spouseID. ID of the spouse
-##'   \item sex.  Sex of the ID: 1 is male; 0 is female
+##'   \item \code{personID}.  Person identification variable
+##'   \item \code{famID}. Family identification variable
+##'   \item \code{name}.  Name of the person
+##'   \item \code{gen}. Generation of the person
+##'   \item \code{momID}.  ID of the mother
+##'   \item \code{dadID}.  ID of the father
+##'   \item \code{spouseID}. ID of the spouse
+##'   \item \code{sex}.  Sex of the ID: 1 is male; 0 is female
 ##'
 ##' }
 ##'