+++ NEWS.md
@@ -1,20 +1,23 @@
 # BGmisc 1.2.0
+
 * Added numerous code checks, increased code coverage to 85%
 * Replaced sapply usage
 * Added additional data validation checks
 * Accompanying paper published in the Journal of Open Source Software
 
 # BGmisc 1.1.0
+
 * Added ability to simulate twins
 * Can now trace paternal and maternal lines
 * There's now a Harry Potter pedigree
 
 # BGmisc 1.0.1
+
 * Hot fix to resolve plotPedigree wrapper function breaking for pedigrees that contained multiple families
 
 # BGmisc 1.0
 
-* Added major update to include simulations, plotting, and examples. 
+* Added major update to include simulations, plotting, and examples.
 
 # BGmisc 0.1